requirements.txt
Django
djangorestframework
PyYAML
requests
django-cors-headers
